local storage persists user's room and theme settings
