+++ client/public/settings.json
@@ -1,0 +1,6 @@
+{
+  "displayName": "",
+  "hideRealName": true,
+  "currentRoom": "central",
+  "currentTheme": "Mono Ocean"
+}
